Add promises to CartItems componentand add popcorn object to cart items model and on model build create a many popcorn to one cart relationship

+++ MomAndPopShop/Models/Popcorn.cs
@@ -20,12 +20,12 @@ namespace MomAndPopShop.Models
         [Required(ErrorMessage = "Price is required")]
         [Range(0.01, double.MaxValue, ErrorMessage = "Price must be 0.01 or greater.")]
         [Precision(18, 2)]
-        public decimal? PopcornPrice { get; set; }
+        public double? PopcornPrice { get; set; }
 
         [Range(0.01, int.MaxValue, ErrorMessage = "Quantity must be 1 or greater.")]
         public int? Quantity { get; set; }
 
-        public Popcorn(string name, string description, decimal? popcornPrice)
+        public Popcorn(string name, string description, double? popcornPrice)
         {
             Name = name;
             Description = description;
